+++ front/src/components/ShareModal/KakaoButton/index.tsx
@@ -1,57 +1,26 @@
 import React, { useEffect } from 'react';
 import styled from 'styled-components';
+import kakaoLogo from '../../../images/kakao.png';
 
 const KakaoButton = (): JSX.Element => {
-  useEffect(() => {
-    createButton();
-  }, []);
   const createButton = () => {
     if (window.Kakao) {
       const kakao = window.Kakao;
       if (!kakao.isInitialized()) kakao.init(process.env.REACT_APP_KAKAO_KEY);
-      kakao.Link.createDefaultButton({
-        // Render 부분 id=kakao-link-btn 을 찾아 그부분에 렌더링을 합니다
-        container: '#kakao-link-btn',
-        objectType: 'feed',
-        content: {
-          title: '타이틀',
-          description: '#리액트 #카카오 #공유버튼',
-          imageUrl: 'IMAGE_URL', // i.e. process.env.FETCH_URL + '/logo.png'
-          link: {
-            mobileWebUrl: window.location.href,
-            webUrl: window.location.href,
-          },
-        },
-        social: {
-          likeCount: 77,
-          commentCount: 55,
-          sharedCount: 333,
-        },
-        buttons: [
-          {
-            title: '웹으로 보기',
-            link: {
-              mobileWebUrl: window.location.href,
-              webUrl: window.location.href,
-            },
-          },
-          {
-            title: '앱으로 보기',
-            link: {
-              mobileWebUrl: window.location.href,
-              webUrl: window.location.href,
-            },
-          },
-        ],
-      });
     }
   };
+  useEffect(() => {
+    createButton();
+  }, []);
+  const handleButton = () => {
+    window.Kakao.Link.sendScrap({
+      requestUrl: window.location.href,
+    });
+  };
   return (
-    <div className="kakao-share-button">
-      <button type="button" id="kakao-link-btn">
-        <img src="/icons/kakao.png" alt="kakao-share-icon" />
-      </button>
-    </div>
+    <button type="button" onClick={handleButton}>
+      <img src={kakaoLogo} alt="kakao-logo" />
+    </button>
   );
 };
