+++ front/src/pages/Main/index.tsx
@@ -33,6 +33,15 @@ function Main(): JSX.Element {
     }
   };
   useEffect(() => {
+    const script = document.createElement('script');
+    script.src = 'https://developers.kakao.com/sdk/js/kakao.js';
+    script.async = true;
+    document.body.appendChild(script);
+    return () => {
+      document.body.removeChild(script);
+    };
+  }, []);
+  useEffect(() => {
     getUserInfo();
   }, []);
   return (
